requirements.txt
PyYAML
torch
ultralytics
